# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}